+++ mycroft/client/enclosure/__init__.py
@@ -280,10 +280,10 @@ class Mark1Enclosure(Enclosure):
         initiate_display_manager_ws()
 
     def on_arduino_responded(self, event=None):
-        from mycroft.enclosure.mark1.arduino import EnclosureArduino
-        from mycroft.enclosure.mark1.eyes import EnclosureEyes
-        from mycroft.enclosure.mark1.mouth import EnclosureMouth
-        from mycroft.enclosure.mark1.weather import EnclosureWeather
+        from mycroft.client.enclosure.arduino import EnclosureArduino
+        from mycroft.client.enclosure.eyes import EnclosureEyes
+        from mycroft.client.enclosure.mouth import EnclosureMouth
+        from mycroft.client.enclosure.weather import EnclosureWeather
         self.eyes = EnclosureEyes(self.ws, self.writer)
         self.mouth = EnclosureMouth(self.ws, self.writer)
         self.system = EnclosureArduino(self.ws, self.writer)
