+++ mycroft/client/enclosure/api.py
@@ -1,0 +1,1 @@
+from mycroft.enclosure.api import *
