+++ mycroft/client/enclosure/arduino.py
