+++ mycroft/client/enclosure/display_manager.py
@@ -1,0 +1,1 @@
+from mycroft.enclosure.display_manager import *
