+++ mycroft/client/enclosure/eyes.py
