+++ mycroft/client/enclosure/mouth.py
