+++ mycroft/client/enclosure/version.txt
