+++ mycroft/client/enclosure/weather.py
